+++ yarn-project/p2p/src/services/tx_collection/index.ts
@@ -1,4 +1,3 @@
-export { TxCollection, type FastCollectionRequestInput } from './tx_collection.js';
-export { type IReqRespTxsCollector } from './fast_tx_collection.js';
+export { TxCollection, type FastCollectionRequestInput, type IReqRespTxsCollector } from './tx_collection.js';
 export { type TxSource, createNodeRpcTxSources, NodeRpcTxSource } from './tx_source.js';
 export { FileStoreTxSource, createFileStoreTxSources } from './file_store_tx_source.js';
